backend/requirements.txt
boto3>=1.26.0
requests
parallel-web
